refactor code and re-arrange folder
