+++ picketeditor/src/main/java/com/skrumble/picketeditor/editor/video/VideoTrimmerUtil.java
@@ -11,10 +11,10 @@ import com.github.hiteshsondhi88.libffmpeg.FFmpeg;
 import com.github.hiteshsondhi88.libffmpeg.LoadBinaryResponseHandler;
 import com.github.hiteshsondhi88.libffmpeg.exceptions.FFmpegCommandAlreadyRunningException;
 import com.github.hiteshsondhi88.libffmpeg.exceptions.FFmpegNotSupportedException;
-import com.skrumble.picketeditor.BackgroundExecutor;
-import com.skrumble.picketeditor.editor.video.public_interface.OnCompletion;
-import com.skrumble.picketeditor.editor.video.public_interface.VideoTrimListener;
-import com.skrumble.picketeditor.picker.utility.Utility;
+import com.skrumble.picketeditor.executor.BackgroundExecutor;
+import com.skrumble.picketeditor.public_interface.OnCompletion;
+import com.skrumble.picketeditor.public_interface.VideoTrimListener;
+import com.skrumble.picketeditor.utility.Utility;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
