+++ picketeditor/src/main/java/com/skrumble/picketeditor/executor/BackgroundExecutor.java
@@ -1,4 +1,4 @@
-package com.skrumble.picketeditor;
+package com.skrumble.picketeditor.executor;
 
 import android.util.Log;
