+++ picketeditor/src/main/java/com/skrumble/picketeditor/public_interface/OnCompletion.java
@@ -1,4 +1,4 @@
-package com.skrumble.picketeditor.editor.video.public_interface;
+package com.skrumble.picketeditor.public_interface;
 
 public interface OnCompletion<T, K> {
     void onCompleted(T t, K k);
